requirements.txt
lxml==5.4.0
pyserial==3.5
